refactor: divide store module
