+++ src/client/components/pages/formTest/index.tsx
@@ -1,7 +1,7 @@
 /**
  * import node_modules
  */
-import React, { useEffect } from "react"
+import React from "react"
 import { NextPage } from "next"
 import Head from "next/head"
 import { Form } from "antd"
@@ -25,11 +25,7 @@ interface FormTestProps extends HandleActions {
 }
 
 const FormTest: NextPage<FormTestProps, GetInitialPropsReturn> = props => {
-  const { handleAddField, handleChangeValue, handleRemoveField, fields, handleGetTodoList } = props
-
-  useEffect(() => {
-    handleGetTodoList()
-  }, [])
+  const { handleAddField, handleChangeValue, handleRemoveField, fields } = props
 
   return (
     <Form>
