+++ src/client/store/modules/page/formTest/types.ts
@@ -16,9 +16,6 @@ export interface Field {
 
 export interface InitialState {
   fields: Field[]
-  master: {
-    isLoading: boolean
-  }
 }
 
 export interface AddFieldPayload {
@@ -57,5 +54,4 @@ export interface HandleActions {
   handleAddField: HandleAddField
   handleChangeValue: HandleChangeValue
   handleRemoveField: HandleRemoveField
-  handleGetTodoList: () => void
 }
