+++ src/client/store/modules/page/todo/types.ts
@@ -1,0 +1,12 @@
+/**
+ * main
+ */
+export interface InitialState {
+  master: {
+    isLoading: boolean
+  }
+}
+
+export interface HandleActions {
+  handleGetTodoList: () => void
+}
